app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.8.2
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-asyncio==0.23.8
python-multipart==0.0.9
playwright==1.45.0
jinja2==3.1.4
